style: update fmt config

+++ src/Counter.sol
@@ -2,9 +2,9 @@
 pragma solidity ^0.8.16;
 
 contract Counter {
-  uint public number;
+  uint256 public number;
 
-  function setNumber(uint newNumber) public {
+  function setNumber(uint256 newNumber) public {
     number = newNumber;
   }
